♻ Rename to getMakerTakerOrders, 🔥 Remove unused composed method
