+++ lib/order/structure/index.js
@@ -1,7 +1,7 @@
 const structure = require('./structure');
 
 structure.getTakerOrders = require('./getTakerOrders');
-structure.getMakerTakerTxns = require('./getMakerTakerTxns');
+structure.getMakerTakerTxns = require('./getMakerTakerOrders');
 structure.withMakerTxns = require('./maker/withMakerTxns');
 
 module.exports = structure;
