+++ lib/order/structure/structure.js
@@ -2,7 +2,7 @@
 const algosdk = require('algosdk');
 const withCloseAlgoTxns = require('../txns/close/withCloseAlgoTxns');
 const withCloseAssetTxns = require('../txns/close/withCloseAssetTxns');
-const getMakerTakerTxns = require('./getMakerTakerTxns');
+const getMakerTakerTxns = require('./getMakerTakerOrders');
 const withMakerTxns = require('./maker/withMakerTxns');
 const withExecuteTxns = require('./taker/withExecuteTxns');
 const getTakerOrders = require('./getTakerOrders');
