--- lib/order/structure/withMakerTakerTxns.js
@@ -1,20 +1,0 @@
-const getMakerTakerTxns = require('./getMakerTakerTxns');
-
-/**
- * Composed MakerTakerTxns transaction
- * @param {AlgodexApi} api
- * @param {Order} order
- * @return {Promise<*>}
- * @ignore
- */
-async function withMakerTakerTxns(api, order) {
-  return {
-    ...order,
-    contract: {
-      ...order.contract,
-      txns: await getMakerTakerTxns(api, order),
-    },
-  };
-}
-
-module.exports = withMakerTakerTxns;
